[TS-285] Refactor: month chart directory refactoring
